설정 파일, 패키지 구조 정리 및 주석 삭제
