+++ src/main/java/was/example/domain/member/Member.java
@@ -1,0 +1,4 @@
+package was.example.domain.member;
+
+public class Member {
+}
