+++ src/main/java/was/example/domain/partner/PartnerInfo.java
@@ -1,0 +1,25 @@
+package was.example.domain.partner;
+
+import lombok.Getter;
+
+/*
+해당 Class는 Partner Domain에 포함시킬지 고민 필요
+ */
+@Getter
+public class PartnerInfo {
+    private final Long id;
+    private final String partnerToken;
+    private final String partnerName;
+    private final String businessNo;
+    private final String email;
+    private final Partner.Status status;
+
+    public PartnerInfo(Partner partner) {
+        this.id = partner.getId();
+        this.partnerToken = partner.getPartnerToken();
+        this.partnerName = partner.getPartnerName();
+        this.businessNo = partner.getBusinessNo();
+        this.email = partner.getEmail();
+        this.status = partner.getStatus();
+    }
+}
