+++ src/main/java/was/example/infrastructure/PartnerRepository.java
@@ -1,0 +1,10 @@
+package was.example.infrastructure;
+
+import was.example.domain.partner.Partner;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PartnerRepository extends JpaRepository<Partner, Long> {
+    Optional<Partner> findByPartnerToken(String partnerToken);
+}
