--- src/main/java/was/examplePakege/example/domain/member/Member.java
@@ -1,4 +1,0 @@
-package was.examplePakege.example.domain.member;
-
-public class Member {
-}
