--- src/main/java/was/examplePakege/example/domain/partner/Partner.java
@@ -1,61 +1,0 @@
-package was.examplePakege.example.domain.partner;
-
-import was.common.exception.InvalidParamException;
-import was.common.util.TokenGenerator;
-import was.common.AbstractEntity;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-
-import javax.persistence.*;
-
-@Slf4j
-@Getter
-@Entity
-@NoArgsConstructor
-@Table(name = "partners")
-public class Partner extends AbstractEntity {
-    private static final String PREFIX_PARTNER = "ptn_";
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String partnerToken;
-    private String partnerName;
-    private String businessNo;
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    private Status status;
-
-    @Getter
-    @RequiredArgsConstructor
-    public enum Status {
-        ENABLE("활성화"), DISABLE("비활성화");
-        private final String description;
-    }
-
-    @Builder
-    public Partner(String partnerName, String businessNo, String email) {
-        if (StringUtils.isEmpty(partnerName)) throw new InvalidParamException("empty partnerName");
-        if (StringUtils.isEmpty(businessNo)) throw new InvalidParamException("empty businessNo");
-        if (StringUtils.isEmpty(email)) throw new InvalidParamException("empty email");
-
-        this.partnerToken = TokenGenerator.randomCharacterWithPrefix(PREFIX_PARTNER);
-        this.partnerName = partnerName;
-        this.businessNo = businessNo;
-        this.email = email;
-        this.status = Status.ENABLE;
-    }
-
-    public void enable() {
-        this.status = Status.ENABLE;
-    }
-
-    public void disable() {
-        this.status = Status.DISABLE;
-    }
-}
