--- src/main/java/was/examplePakege/example/domain/partner/interfaces/PartnerReader.java
@@ -1,8 +1,0 @@
-package was.examplePakege.example.domain.partner.interfaces;
-
-import was.examplePakege.example.domain.partner.Partner;
-
-public interface PartnerReader {
-    Partner getPartner(Long partnerId);
-    Partner getPartner(String partnerToken);
-}
