--- src/main/java/was/examplePakege/example/domain/partner/interfaces/PartnerService.java
@@ -1,8 +1,0 @@
-package was.examplePakege.example.domain.partner.interfaces;
-
-import was.examplePakege.example.domain.partner.PartnerCommand;
-import was.examplePakege.example.domain.partner.PartnerInfo;
-
-public interface PartnerService {
-    PartnerInfo registerPartner(PartnerCommand command);
-}
