--- src/main/java/was/examplePakege/example/infrastructure/PartnerRepository.java
@@ -1,10 +1,0 @@
-package was.examplePakege.example.infrastructure;
-
-import was.examplePakege.example.domain.partner.Partner;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PartnerRepository extends JpaRepository<Partner, Long> {
-    Optional<Partner> findByPartnerToken(String partnerToken);
-}
