--- src/main/java/was/examplePakege/example/infrastructure/Read/PartnerReadImpl.java
@@ -1,28 +1,0 @@
-package was.examplePakege.example.infrastructure.Read;
-
-import was.common.exception.EntityNotFoundException;
-import was.examplePakege.example.domain.partner.Partner;
-import was.examplePakege.example.domain.partner.interfaces.PartnerReader;
-import was.examplePakege.example.infrastructure.PartnerRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class PartnerReadImpl implements PartnerReader {
-    private final PartnerRepository partnerRepository;
-
-    @Override
-    public Partner getPartner(Long partnerId) {
-        return partnerRepository.findById(partnerId)
-                .orElseThrow(EntityNotFoundException::new);
-    }
-
-    @Override
-    public Partner getPartner(String partnerToken) {
-        return partnerRepository.findByPartnerToken(partnerToken)
-                .orElseThrow(EntityNotFoundException::new);
-    }
-}
