--- src/main/java/was/examplePakege/example/infrastructure/Store/PartnerStoreImpl.java
@@ -1,29 +1,0 @@
-package was.examplePakege.example.infrastructure.Store;
-
-import was.common.exception.InvalidParamException;
-import was.examplePakege.example.domain.partner.Partner;
-import was.examplePakege.example.domain.partner.interfaces.PartnerStore;
-import was.examplePakege.example.infrastructure.PartnerRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class PartnerStoreImpl implements PartnerStore {
-
-    private final PartnerRepository partnerRepository;
-
-    @Override
-    public Partner store(Partner partner) {
-        if (StringUtils.isEmpty(partner.getPartnerToken())) throw new InvalidParamException("partner.getPartnerToken()");
-        if (StringUtils.isEmpty(partner.getPartnerName())) throw new InvalidParamException("partner.getPartnerName()");
-        if (StringUtils.isEmpty(partner.getBusinessNo())) throw new InvalidParamException("partner.getBusinessNo()");
-        if (StringUtils.isEmpty(partner.getEmail())) throw new InvalidParamException("partner.getEmail()");
-        if (partner.getStatus() == null) throw new InvalidParamException("partner.getStatus()");
-
-        return partnerRepository.save(partner);
-    }
-}
