--- src/main/java/was/examplePakege/example/interfaces/response/RegisterResponse.java
@@ -1,25 +1,0 @@
-package was.examplePakege.example.interfaces.response;
-
-import lombok.Getter;
-import lombok.ToString;
-import was.examplePakege.example.domain.partner.Partner;
-import was.examplePakege.example.domain.partner.PartnerInfo;
-
-@Getter
-@ToString
-public class RegisterResponse {
-
-    private final String partnerToken;
-    private final String partnerName;
-    private final String businessNo;
-    private final String email;
-    private final Partner.Status status;
-
-    public RegisterResponse(PartnerInfo partnerInfo) {
-            this.partnerToken = partnerInfo.getPartnerToken();
-            this.partnerName = partnerInfo.getPartnerName();
-            this.businessNo = partnerInfo.getBusinessNo();
-            this.email = partnerInfo.getEmail();
-            this.status = partnerInfo.getStatus();
-    }
-}
