--- src/main/java/was/examplePakege/examplenotification/domain/NotificationService.java
@@ -1,5 +1,0 @@
-package was.examplePakege.examplenotification.domain;
-
-public interface NotificationService {
-    void sendEmail(String email, String title, String description);
-}
