--- src/main/java/was/examplePakege/examplenotification/infrastructure/NotificationExecutor.java
@@ -1,15 +1,0 @@
-package was.examplePakege.examplenotification.infrastructure;
-
-import was.examplePakege.examplenotification.domain.NotificationService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-public class NotificationExecutor implements NotificationService {
-
-    @Override
-    public void sendEmail(String email, String title, String description) {
-        log.info("sendEmail");
-    }
-}
